Updated fragments with layout & Functionality from dev
