+++ app/src/main/java/com/eventplanner/fragments/solutions/TopSolutionsFragment.java
@@ -1,6 +1,5 @@
 package com.eventplanner.fragments.solutions;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,10 +9,10 @@ import android.widget.ListView;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
 
 import com.eventplanner.R;
-import com.eventplanner.activities.solutions.AllProductsActivity;
-import com.eventplanner.activities.solutions.AllServicesActivity;
 import com.eventplanner.adapters.solutions.SolutionListAdapter;
 import com.eventplanner.model.solutions.Product;
 import com.eventplanner.model.solutions.ReservationType;
@@ -56,15 +55,15 @@ public class TopSolutionsFragment extends Fragment {
         SolutionListAdapter adapter = new SolutionListAdapter(getContext(), solutions);
         listView.setAdapter(adapter);
 
+        NavController navController = Navigation.findNavController(getActivity(), R.id.fragment_nav_content_main);
+
         Button browseServicesButton = rootView.findViewById(R.id.browse_services_button);
         browseServicesButton.setOnClickListener(v -> {
-            Intent intent = new Intent(getActivity(), AllServicesActivity.class);
-            startActivity(intent);
+            navController.navigate(R.id.nav_all_services);
         });
         Button browseProductsButton = rootView.findViewById(R.id.browse_products_button);
         browseProductsButton.setOnClickListener(v -> {
-            Intent intent = new Intent(getActivity(), AllProductsActivity.class);
-            startActivity(intent);
+            navController.navigate(R.id.nav_all_products);
         });
 
         return rootView;
