--- app/src/main/res/layout/activity_all_services.xml
@@ -1,102 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:id="@+id/all_events_textview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/all_services"
-        android:textSize="40sp" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:padding="16dp">
-
-        <androidx.appcompat.widget.SearchView
-            android:id="@+id/search_bar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:queryHint="Search services"
-            android:iconifiedByDefault="false"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:background="@drawable/rounded_search_view"
-            android:radius="20dp"
-            android:layout_margin="5dp"/>
-
-        <LinearLayout
-            android:id="@+id/linear_buttons"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/button_filter_services"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:layout_marginTop="20dp"
-                android:layout_marginBottom="10dp"
-                android:layout_marginStart="5dp"
-                app:backgroundTint="@color/cool_purple"
-                android:background="@drawable/button_rounded"
-                android:text="@string/filter" />
-        </LinearLayout>
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_gravity="center"
-        android:padding="15dp">
-
-        <Button
-            android:id="@+id/button2"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:text="@string/previous_text"
-            android:layout_marginEnd="8dp"
-            />
-        <Button
-            android:id="@+id/button3"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:text="1"
-            android:layout_marginEnd="8dp"/>
-        <Button
-            android:id="@+id/button4"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:text="2"
-            android:layout_marginEnd="8dp"/>
-        <Button
-            android:id="@+id/button5"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:text="3"
-            android:layout_marginEnd="8dp"/>
-        <Button
-            android:id="@+id/button6"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:text="@string/next_text"
-            android:layout_marginEnd="8dp"/>
-    </LinearLayout>
-
-    <androidx.fragment.app.FragmentContainerView
-        android:name="com.eventplanner.fragments.solutions.AllServicesFragment"
-        android:id="@+id/fragment_all"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-
-</LinearLayout>
